# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "一款面向小型团队的任务进度共享软件",
        "title": "Team",
        "termsOfService": "http://swagger.io/terrms",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:9918",
    "paths": {
        "/info/donelist": {
            "get": {
                "description": "“获取已完成任务”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasklist"
                ],
                "summary": "“已完成任务”",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "获取失败"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/info/donetask": {
            "put": {
                "description": "“取消对钩”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasklist"
                ],
                "summary": "“取消任务的完成”",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/info/donetask/:task_id": {
            "put": {
                "description": "“打对钩完成任务”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasklist"
                ],
                "summary": "“完成任务”",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "任务id",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "任务完成"
                    },
                    "401": {
                        "description": "验证失败"
                    }
                }
            }
        },
        "/info/todolist": {
            "get": {
                "description": "“获取未完成任务”",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasklist"
                ],
                "summary": "“任务待办”",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "获取失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
...
        "/team": {
            "get": {
                "description": "\"刚登陆后的第一个界面\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "\"查看用户加入的所有团队\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "获取失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            },
            "post": {
                "description": "\"创建一个新的团队\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "\"创建团队\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "teamInfo",
                        "name": "teamInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "团队创建成功"
                    },
                    "400": {
                        "description": "创建失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/team/:team_id": {
            "get": {
                "description": "\"单击团队名查看团队信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "\"查看团队\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team_id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "获取失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
        "/team/paticipation": {
            "post": {
                "description": "\"加入一个新的团队\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "\"加入团队\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "team",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserTeam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "加入成功"
                    },
                    "400": {
                        "description": "加入失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Team": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "creator_id": {
                    "type": "integer"
                },
                "team_coding": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "teamname": {
                    "type": "string"
                }
            }
        },
...
        "model.UserTeam": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "team_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}